[swonseok] fix(db): care_items.core/base 를 DECIMAL(3,1) 로
마스터 시드에 소수가 9건 있습니다.
  core 2.5 (2건)
  base 0.8 1.2 1.4 1.6 2.2 2.5 3.5 (각 1건)

SMALLINT 에 넣으면 반올림되고, load = base + FQL 이 어긋나
score 가 약 0.55 움직입니다. 임계값이 0.8 과 -2.5 라서 판정이 뒤집힙니다.
SubtractPipelineCheck 31건이 통과한 결과와 실서버 결과가 달라집니다.

FLOAT/DOUBLE 은 쓰지 않습니다 - 점수식에 오차가 두 번 들어갑니다.
DECIMAL(3,1) 은 -99.9~99.9 이고 현재 값은 0.8~5 입니다.

core/base 는 care_items 한 곳뿐입니다. eval_results 에는 floor 스냅숏만
있고 core/base 스냅숏은 없습니다.

예외 허용 근거: DISCREPANCIES-0819 E-6 확대, N-1

+++ docs/schema_v63.sql
@@ -138,8 +138,12 @@ CREATE TABLE care_items (                                         -- 32행
     name                VARCHAR(255) NOT NULL,
     floor               VARCHAR(20)  NOT NULL,
     evidence_level      VARCHAR(20)  NOT NULL,
-    core                SMALLINT     NOT NULL,                    -- 중요도. 점수식 입력
-    base                SMALLINT     NOT NULL,                    -- 기본 부담. 점수식 입력
+    -- 2026-08-20 SMALLINT -> DECIMAL(3,1). 마스터 시드에 소수가 9건 있습니다
+    -- (core 2.5 2건 · base 0.8 1.2 1.4 1.6 2.2 2.5 3.5 각 1건).
+    -- 반올림되면 load 가 어긋나 임계값 0.8 · -2.5 에서 판정이 뒤집힙니다.
+    -- FLOAT/DOUBLE 은 쓰지 않습니다 — 점수식에 오차가 두 번 들어갑니다.
+    core                DECIMAL(3,1) NOT NULL,                    -- 중요도. 점수식 입력
+    base                DECIMAL(3,1) NOT NULL,                    -- 기본 부담. 점수식 입력
     minutes             SMALLINT     NOT NULL,                    -- 고정 소요 시간(분)
     frequency_editable  BOOLEAN      NOT NULL,                    -- false 면 빈도 UI 를 띄우지 않습니다
     default_frequency   VARCHAR(20)  NULL,
